requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.8
django-filter==25.2
django-jazzmin==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
openapi-codec==1.3.2
pillow==12.0.0
PyJWT==2.10.1
requests==2.32.5
simplejson==3.20.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
